src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
azure-core==1.25.1
azure-cosmos==4.3.0
azure-identity==1.11.0
azure_storage_blob==12.14.1
azure-monitor-opentelemetry==1.0.0b13
opentelemetry-instrumentation-fastapi
beanie==1.16.6
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
fastapi==0.95.1
fastapi-code-generator==0.3.5
filelock==3.8.0
frozenlist==1.3.1
google-api-core==2.10.1
googleapis-common-protos==1.56.4
h11==0.13.0
idna==3.4
iniconfig==1.1.1
install==1.3.5
isodate==0.6.1
json2yaml==1.2.0
mongomock-motor==0.0.12
motor==3.0.0
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
protobuf==4.21.6
psutil==5.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.2.0
python-jose[cryptography]==3.3.0
pyjwt==2.7.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-mock==3.9.0
pytest-sugar==0.9.5
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli==2.0.1
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
pylint==2.15.5
uvicorn==0.18.3
jinja2==3.1.2
GitPython==3.1.31
retry==0.9.2
deepdiff[murmur]==6.3.0
pyyaml==6.0
azure-devops==7.1.0b3
